Renamed mdp module to markov; Renamed state_space and action_space to states and actions; Updated test script to ask user whether or not he wants to run slow tests; Added ndarray_index_from function to yaaf; Bumped version to 1.0.4;

+++ test.sh
@@ -5,21 +5,30 @@
 # ############### #
 
 while true; do
-    read -p "Do you wish to setup a new virtual environment for testing YAAF? (Y/N) > " answer
-    case $answer in
+    read -p "Do you wish to setup a new virtual environment for testing YAAF? (Y/N) > " answervenv
+    case $answervenv in
         [Yy]* ) break ;;
         [Nn]* ) break ;;
         * ) echo "Please answer [y] or [n].";;
     esac
 done
 
-if [ "$answer" == "y" ]; then
+if [ "$answervenv" == "y" ]; then
   python -m venv yaaf_test_environment
   source yaaf_test_environment/bin/activate
   pip install --upgrade pip setuptools
   pip install -r requirements.txt
   pip install -e .
 fi
+
+while true; do
+    read -p "Do you wish to skip supervised learning tests? (slow) (Y/N) > " answerskipslow
+    case $answerskipslow in
+        [Yy]* ) break ;;
+        [Nn]* ) break ;;
+        * ) echo "Please answer [y] or [n].";;
+    esac
+done
 
 # ######### #
 # Run tests #
@@ -28,21 +37,23 @@ fi
 echo "Running framework tests (. = passed, E = Error)"
 python -m unittest discover -s tests -t tests
 
-echo "Running supervised learning character recognition example tests (slow) (. = passed, E = Error)"
-cp -r examples/character_recognition/ocr_dataset .
-python -m unittest discover -s examples/character_recognition -t examples/character_recognition
-rm -rf ./ocr_dataset
-
 echo "Running pursuit example tests (. = passed, E = Error)"
 cp -r examples/pursuit/pretrained-agents .
 python -m unittest discover -s examples/pursuit -t examples/pursuit
 rm -rf ./pretrained-agents
 
+if [ "$answerskipslow" == "n" ]; then
+  echo "Running supervised learning character recognition example tests (slow) (. = passed, E = Error)"
+  cp -r examples/character_recognition/ocr_dataset .
+  python -m unittest discover -s examples/character_recognition -t examples/character_recognition
+  rm -rf ./ocr_dataset
+fi
+
 # ####### #
 # Cleanup #
 # ####### #
 
-if [ "$answer" == "y" ]; then
+if [ "$answervenv" == "y" ]; then
   source deactivate
   rm -rf yaaf_test_environment
 fi
